requirements.txt
strands-agents
strands-agents-tools
httpx
wikipedia
pydantic
python-dotenv
psycopg2-binary
sshtunnel
